:hammer_and_wrench: Travis CI improvements
This commit gets Travis to successfully test the repo again. This was
done in coordination with some changes to the upstream template to make
it fit my use case better for how I maintain all of these docs.

I also move the Travis tests to a job matrix, so it's easier to add
multiple tests into an existing Travis test suite if it already exists.

Signed-off-by: Justin W. Flory <git@jwf.io>